refactor: 클래스의 디렉토리 위치 수정
