+++ backend/src/test/java/backend/mulkkam/support/MemberFixture.java
@@ -4,7 +4,7 @@ import backend.mulkkam.intake.domain.vo.Amount;
 import backend.mulkkam.member.domain.Member;
 import backend.mulkkam.member.domain.vo.Gender;
 import backend.mulkkam.member.domain.vo.MemberNickname;
-import backend.mulkkam.member.dto.PhysicalAttributes;
+import backend.mulkkam.member.domain.vo.PhysicalAttributes;
 
 public class MemberFixture {
